Fixes #22 - drastically reduce the output by getting rid of a line that should be on testing branch only, and suppress all output from the kindleunpack script (with some sys.stderr/stdout manipulation)
For the record/my memory: all the output from kindleunpack isn't directly interersting, we just wanted the dumped files.  The right solution would be to control that output with a flag, but I've gone for the smallest fix possible - which will help if I choose to upgrade the kindleunpack version we have later.