Fixes #25, mostly.  The problem here was related to encodings.  XRay data is interested in addresses in bytes, but if we have a string in an encoding where a character can be longer than a byte (unicode, etc.) then things will fail - this explains why we were seeing locations being offset by two (as a single 3-byte unicode character was being counted as a length of 1).
The problem is almost entirely fixed, but there are a very small number of still broken cases and I can't see entirely why.  In one case we are overshooting the actual location (in all previous cases we'd either get the address right or would undershoot, for the reasons explained above).  I might look in to this later but it is sufficiently rare to not be a major issue - and if I change the HTML parsing code enough then it might go or be replaced by new and exciting bugs.

Fixes #26 - by default no logging happens, but if a logfile is specified then any logfile.write*() calls in the code will cause logging to occur.