Move XRayLogfile to utility module
Move all the X-Ray old format (JSON) generating code to the function that is called to output it - it is all specific to generating JSON with the exception of one useful warning I've now duplicated in the SQLite path too.  Also renamed the functions to make clear they aren't (intended to be) called externally.